docker/minimal-examples/safe-api-key-docker-compose/pyproject.toml
[project]
name = "safe-api-key-docker-compose"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    "langchain>=0.3.26",
    "langchain-openai>=0.3.25",
    "python-dotenv>=1.1.1",
]
